fix(alder-region): widen Position line/column to u32
A line longer than 65535 bytes or a file with more than 65535 lines
overflowed the u16 row/column counters (a debug panic in advance(), a
silent wrap in release that misplaced every later region and error).
Position, Row and Col are u32 now; advance() stays plain arithmetic.
Regression tests parse a 70_000-byte line and a 70_000-line module.

+++ crates/alder-parse/src/lib.rs
@@ -31,8 +31,8 @@ mod type_;
 // error renderer in another crate can name them.
 pub use keyword::{Keyword, SqlWord};
 
-pub type Row = u16;
-pub type Col = u16;
+pub type Row = u32;
+pub type Col = u32;
 
 /// Saved parser state for backtracking.
 #[derive(Clone, Copy)]
